5. Переместил список контактов в фейковый репозиторий
