+++ app/src/main/java/ru/ds/gitapp/ui/main/MainFragment.kt
@@ -1,4 +1,4 @@
-package ru.ds.gitapp.ui
+package ru.ds.gitapp.ui.main
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,8 +6,8 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
-import ru.ds.gitapp.ProfileAdapter
 import ru.ds.gitapp.R
+import ru.ds.gitapp.data.UserRepositoryImp
 import ru.ds.gitapp.databinding.MainFragmentBinding
 import ru.ds.gitapp.domain.Profile
 
@@ -20,13 +20,7 @@ class MainFragment : Fragment() {
     private var adapter: ProfileAdapter? = null
 
 
-    val userList = listOf(
-        Profile("UserName1","email@example.com", R.drawable.ic_profile_icon1),
-        Profile("UserName2","email@example.com", R.drawable.ic_profile_icon2),
-        Profile("UserName3","email@example.com", R.drawable.ic_profile_icon3),
-        Profile("UserName4","email@example.com", R.drawable.ic_profile_icon4),
-
-    )
+   private val userLocalList = UserRepositoryImp()
 
 
     override fun onCreateView(
@@ -47,7 +41,7 @@ class MainFragment : Fragment() {
         binding.apply {
 
             val list = ArrayList<Profile>()
-            list.addAll(userList)
+            list.addAll(userLocalList.getUserFromLocalStorage())
 
             recycleView.hasFixedSize()
             recycleView.layoutManager = LinearLayoutManager(requireContext())
