+++ app/src/main/java/ru/ds/gitapp/ui/main/ProfileAdapter.kt
@@ -1,4 +1,4 @@
-package ru.ds.gitapp
+package ru.ds.gitapp.ui.main
 
 import android.content.Context
 import android.content.Intent
@@ -6,9 +6,9 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
+import ru.ds.gitapp.R
 import ru.ds.gitapp.databinding.ProfileContentBinding
 import ru.ds.gitapp.domain.Profile
-import ru.ds.gitapp.ui.ContentActivity
 
 class ProfileAdapter(listArray: ArrayList<Profile>, context: Context): RecyclerView.Adapter<ProfileAdapter.SectorsHolder>() {
